+++ CHANGELOG.md
@@ -15,6 +15,7 @@ before starting to add changes. Use example [placed in the end of the page](#exa
   Encrypts all elements if encryption enabled.
 - [#114](https://github.com/OS2Forms/os2forms/pull/114)
   Encrypted computed elements.
+- [OS-74] Updating DAWA matrikula select with Datafordeler select
 
 ## [3.15.3] 2024-06-25